_legacy/v3/requirements.txt
# ============================================================
# Seven AI v2.6 - Advanced Sentience Architecture
# Install: pip install -r requirements.txt
# Also required: Ollama with llama3.2 + llama3.2-vision models
# ============================================================

# --- Core (required) ---
edge-tts>=6.1.0            # Natural neural TTS (Microsoft voices, free)
pygame>=2.5.0              # Audio playback for edge-tts with instant interrupt
pyttsx3>=2.90              # Fallback text-to-speech (offline, robotic)
vosk>=0.3.45               # Offline speech recognition (replaces online-only SpeechRecognition)
SpeechRecognition>=3.10.0  # Voice recognition (fallback / online)
pyaudio>=0.2.13            # Microphone input for voice barge-in + speech recognition
# Note: PyAudio can be tricky on Windows. If pip fails:
#   pip install pipwin && pipwin install pyaudio
requests>=2.31.0           # HTTP / Ollama API
psutil>=5.9.0              # Process management, performance monitoring
schedule>=1.2.0            # Background task scheduling
python-dateutil>=2.8.2     # Date/time parsing
networkx>=3.1              # Knowledge graph

# --- Web Search & Content ---
googlesearch-python>=1.2.3 # Google search
beautifulsoup4>=4.12.0     # Web page content extraction

# --- Vision System ---
opencv-python>=4.8.0       # Camera / vision (webcam + IP cameras via OpenCV)
numpy>=1.24.0              # Image processing (installed with opencv)
Pillow>=10.0.0             # Screenshot capture + image manipulation
# Also requires: ollama pull llama3.2-vision (for scene understanding)

# --- Screen Control ---
pyautogui>=0.9.54          # Mouse/keyboard automation, screenshots
# Note: pyautogui.FAILSAFE=True — move mouse to corner to abort

# --- Keyboard Interrupt (recommended) ---
keyboard>=0.13.5           # ESC/SPACE interrupt while Seven speaks
# Note: On Windows, may require running as administrator

# --- Google Calendar (optional) ---
google-api-python-client>=2.108.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# --- Optional: Vector Memory ---
# Uncomment for semantic memory (large download)
# chromadb>=0.4.22

# --- Optional: Advanced Voice ---
# Uncomment for Whisper speech recognition (large download)
# openai-whisper>=20230314
# faster-whisper>=0.9.0

# --- V3.0: Daemon, API, Scheduler, Multi-Agent ---
fastapi>=0.104.0           # REST API for external control + webhooks
uvicorn>=0.24.0            # ASGI server for FastAPI
apscheduler>=3.10.0        # Persistent task scheduling (SQLite-backed)
sqlalchemy>=2.0.0          # APScheduler job store backend
cryptography>=41.0.0       # Fernet encryption for SSH credential storage
structlog>=23.2.0          # Structured logging (better tracing)

# --- Optional: Vector Memory (ChromaDB) ---
# Uncomment for semantic memory (large download)
# chromadb>=0.4.22

# --- NEAT Neuroevolution (Self-Evolution) ---
neat-python>=0.92          # NEAT algorithm for evolving neural networks

# --- v3.2.15: Web UI (Gradio) ---
gradio>=6.0.0              # Browser-based web interface alongside Tkinter
plotly>=5.0.0              # Interactive knowledge graph visualization

# --- v3.2: Unsurpassable Features ---
pyserial>=3.5              # Hardware embodiment (Arduino/RPi serial)
statsmodels>=0.14.0        # Predictive user modeling (ARIMA time-series)
pandas>=2.0.0              # Data analysis for user predictor

# --- Optional: LoRA Fine-Tuning ---
# Uncomment for real LoRA fine-tuning (large download)
# peft>=0.7.0              # Parameter-Efficient Fine-Tuning
# litgpt>=0.3.0            # Lightweight GPT fine-tuning

# --- Optional: LangGraph Multi-Agent ---
# Uncomment for advanced multi-agent orchestration
# langgraph>=0.0.20

# --- Development ---
# pytest>=8.0.0
# black>=23.10.0
# flake8>=6.1.0

# --- v3.2.17 / v3.2.18: Episodic memory + action item digest ---
# No new deps needed — uses stdlib (sqlite3, json, hashlib)
# openai-whisper is optional for ambient listener (see above)

# --- v3.2.19: MCP server — Seven's brain over Model Context Protocol ---
mcp>=1.0.0                 # Anthropic's official Python SDK (includes FastMCP)
# Note: pydantic is already a transitive dep of fastapi — no pin needed
